Changed db_schema.py to actually use the salts inside the hashing algorithm.
changed app.py to be able to recieve a /login route and be posted a username and password to check against users in the database to log users in.
if users input correct details, generate a web token for the user to allow them to access the rest of the website.